requirements.txt
streamlit
python-dotenv
requests
yt-dlp
langchain
openai
PyPDF2
faiss-cpu
tiktoken
euriai